Stack from [ghstack](https://github.com/ezyang/ghstack):
* **#52577 fix nest openmp performance bug in thnn_conv2d**

Differential Revision: [D27063800](https://our.internmc.facebook.com/intern/diff/D27063800)