@mingfeima 2021-02-22
`at::thnn_conv2d` has very low performance on multi-core CPU when `batch_size=1`.

`at::thnn_conv2d` used the gemm-based algorithm (im2col + gemm), which will have a nested openmp loop: 
- outer loop will parallel on dimension of N (batch_size)
- inner loop will do im2col + gemm per each image, the inner loop is forced to go sequential

This is problematic for **single batch size inference** on a **multi-core** run where available HW resources can not be full utilized. This patch will improve `at::thnn_conv2d` and `at::slow_conv_transpose2d` with #51185, also the 3d counterpart.

For `at::thnn_conv2d`, single batch inference improvements:
- avx512 CPU (20 cores): ~11x
- avx2 CPU (12 cores): ~9x

[TODO] The threading algorithm in `at::parallel_for` introduces load imbalance on some occasions, will port **balance211** algorithm from mkldnn to pytorch in following patches.